Move logger into services
