+++ kiosk/src/js/application_updater.js
@@ -8,7 +8,7 @@ const UserInteractionsManager = require('./user_interactions_manager');
 const NetworkErrorsHandlingService = require('./services/error/network_errors_handling_service');
 const AutoupdateScheduleWatcher = require('./services/data/autoupdate_schedule_watcher');
 const ConnectionStatusService = require('./services/network/connection_status_service');
-const Logger = require('./logger/logger');
+const Logger = require('./services/logger/logger');
 const DataSender = require('./data_sender');
 const DataStorage = require('./storage/data_storage');
 const LocalStorageManager = require('./helpers/local_storage_helper').LocalStorageManager;
