+++ kiosk/src/js/services/error/network_errors_handling_service.js
@@ -1,6 +1,6 @@
 const Subject = require('rxjs').Subject;
 const GlobalErrorsHandlingService = require('./global_errors_handling_service');
-const Logger = require('./../../logger/logger');
+const Logger = require('../logger/logger');
 
 let instance = null;
