+++ kiosk/src/js/services/services_initialiser.js
@@ -3,7 +3,7 @@ class ServicesInitialiser {
 
     //should be called ON application initializing
     static initBaseServices() {
-        require('./../logger/logger').setupLoggerProperties();
+        require('./logger/logger').setupLoggerProperties();
         require('./error/global_errors_handling_service');
         require('./error/network_errors_handling_service');
         require('./error/uncaught_errors_handling_service');
